Remove showdown and jquery as dependencies
Showdown is a dev dependency, as it gets bundled with readout and sent off to dist/

Jquery, while technically a dependency, is so common that I don't think it should be imported and bundled in with readout. Instead it should just fail until a script tag is added referencing it. This is similar to webpack's external dependency feature, and in fact webpack issues a warning of jquery is not available for us.